+++ plugins/phase-distortion-osc/phase-distortion-osc.cpp
@@ -28,7 +28,7 @@
 class PhaseDistortionOscillator : public LV2Plugin {
 public:
   
-  PhaseDistortionOscillator(unsigned long rate, const char*, 
+  PhaseDistortionOscillator(uint32_t rate, const char*, 
                             const LV2_Host_Feature**) 
     : LV2Plugin(4),
       m_osc(rate) {
@@ -41,7 +41,7 @@ public:
   }
 
   
-  void run(unsigned long nframes) {
+  void run(uint32_t nframes) {
     float* f = static_cast<float*>(m_ports[0]);
     float shape = *static_cast<float*>(m_ports[1]);
     float amount = *static_cast<float*>(m_ports[2]);
@@ -53,7 +53,7 @@ public:
     amount = (amount < 0 ? 0 : amount);
     amount = (amount > 0.999 ? 0.999 : amount);
     
-    for (unsigned long i = 0; i < nframes; ++i)
+    for (uint32_t i = 0; i < nframes; ++i)
       output[i] = m_osc.run(440, ishape, amount);
   }
